support for basic auth and api key in combi with cloud id
